# swagger_schema.json
{"openapi":"3.0.0","info":{"version":"1.0.0","title":"Minesweeper","description":"Minesweeper"},"servers":[{"url":"/","description":"Local Dev"}],"paths":{"/boards":{"post":{"tags":["BOARDS"],"description":"Creates a new board!","operationId":"createBoard","parameters":[{"in":"body","name":"rows","schema":{"type":"number","min":3,"max":20,"example":5},"required":false},{"in":"body","name":"columns","schema":{"type":"number","min":3,"max":20,"example":5},"required":false},{"in":"body","name":"difficulty","schema":{"type":"number","example":5,"min":1,"max":100},"required":false}],"responses":{"201":{"description":"The board","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Board"}}}},"500":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["default_error","internal_error"]}}}}}},"503":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["service_unavailable"]}}}}}}}}},"/boards/:board_tag":{"get":{"tags":["BOARDS"],"description":"Gets a saved board!","operationId":"getSavedBoard","responses":{"200":{"description":"The saved board","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Board"}}}},"404":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["board_not_found"]}}}}}},"500":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["default_error","internal_error"]}}}}}},"503":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["service_unavailable","invalid_board"]}}}}}}}}},"/boards/:board_id/save/:board_tag":{"get":{"tags":["BOARDS"],"description":"Saves a board by a tag!","operationId":"saveBoardByTag","responses":{"201":{"description":"Board saved ok"},"400":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["tag_already_exists_error"]}}}}}},"404":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["board_not_found"]}}}}}},"500":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["default_error","internal_error"]}}}}}},"503":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["service_unavailable","invalid_board"]}}}}}}}}},"/boards/:board_id/toggle_flag":{"put":{"tags":["BOARDS"],"description":"Toggles a flag on a position","operationId":"toggleFLag","parameters":[{"in":"params","name":"board_id","schema":{"type":"string"},"required":true},{"in":"body","name":"row","schema":{"type":"number","min":3,"max":20,"example":5},"required":true},{"in":"body","name":"column","schema":{"type":"number","min":3,"max":20,"example":5},"required":true}],"responses":{"200":{"description":"The updated board","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Board"}}}},"400":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["position_has_hint","invalid_position_for_board"]}}}}}},"404":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["board_not_found"]}}}}}},"500":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["default_error","internal_error"]}}}}}},"503":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["service_unavailable","invalid_board"]}}}}}}}}},"/boards/:board_id/reveal":{"put":{"tags":["BOARDS"],"description":"Reveals a position","operationId":"revealPosition","parameters":[{"in":"params","name":"board_id","schema":{"type":"string"},"required":true},{"in":"body","name":"row","schema":{"type":"number","min":3,"max":20,"example":5},"required":true},{"in":"body","name":"column","schema":{"type":"number","min":3,"max":20,"example":5},"required":true}],"responses":{"200":{"description":"The updated board","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Board"}}}},"400":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["position_has_hint","invalid_position_for_board"]}}}}}},"404":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["board_not_found"]}}}}}},"500":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["default_error","internal_error"]}}}}}},"503":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["service_unavailable","invalid_board"]}}}}}}}}},"/boards/:board_id/toggle_question":{"put":{"tags":["BOARDS"],"description":"Toggles a question mark on a position","operationId":"toggleQuestion","parameters":[{"in":"params","name":"board_id","schema":{"type":"string"},"required":true},{"in":"body","name":"row","schema":{"type":"number","min":3,"max":20,"example":5},"required":true},{"in":"body","name":"column","schema":{"type":"number","min":3,"max":20,"example":5},"required":true}],"responses":{"200":{"description":"The updated board","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Board"}}}},"400":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["position_has_hint","invalid_position_for_board"]}}}}}},"404":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["board_not_found"]}}}}}},"500":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["default_error","internal_error"]}}}}}},"503":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["service_unavailable","invalid_board"]}}}}}}}}}},"components":{"schemas":{"Board":{"type":"object","properties":{"id":{"type":"string","format":"uuid","example":"c5613673-f0d5-4ff1-88fe-908fd4640d91"},"rows":{"type":"number","example":1},"columns":{"type":"number","example":1},"difficulty":{"type":"number","example":1},"elapsed_seconds":{"type":"number"},"flags":{"type":"object","example":{"x.y":true}},"questionMarks":{"type":"object","example":{"x.y":true}},"hints":{"type":"object","example":{"x.y":1}}}}},"securitySchemes":{}}}